Se ha hecho el cambio de base de datos de int a double para los costes y los efectos y mejorado la busqueda
- el cambio de int a doble es para permitir decimales y numeros muy grandes, de esta manera se puede conseguir una progresion mucho m,as detallada, controlada y lenta
-las busquedas en la lista de mejoras tienen una progress bar y en principio ya no deberia dar errores de concurrrencia por el sincornized

+++ app/src/main/java/com/example/android_app/AppLifecycle.java
@@ -123,9 +123,9 @@ public class AppLifecycle extends Application implements Application.ActivityLif
             //Log.d("Clicker->", "APPCYCLE -> Timer stopped stopppedfor Game");
         }
 
-        int pcu = scoreManager.getPassiveValue();
-        int acu = scoreManager.getClickValue();
-        int score = scoreManager.getScore();
+        double pcu = scoreManager.getPassiveValue();
+        double acu = scoreManager.getClickValue();
+        double score = scoreManager.getScore();
 
         gameViewModel.updateUserStats(score, pcu, acu);
 
@@ -151,9 +151,9 @@ public class AppLifecycle extends Application implements Application.ActivityLif
         //los datos de llos niveles a los que tiene cada mejora el ussuario se guardan cada vez que compro una mejora
         //control null
         if(activity instanceof Game){
-            int pcu = scoreManager.getPassiveValue();
-            int acu = scoreManager.getClickValue();
-            int score = scoreManager.getScore();
+            double pcu = scoreManager.getPassiveValue();
+            double acu = scoreManager.getClickValue();
+            double score = scoreManager.getScore();
 
             gameViewModel.updateUserStats(score, pcu, acu);
 
@@ -167,9 +167,9 @@ public class AppLifecycle extends Application implements Application.ActivityLif
             editor.apply();
         }
 
-        int pcu = scoreManager.getPassiveValue();
-        int acu = scoreManager.getClickValue();
-        int score = scoreManager.getScore();
+        double pcu = scoreManager.getPassiveValue();
+        double acu = scoreManager.getClickValue();
+        double score = scoreManager.getScore();
 
         gameViewModel.updateUserStats(score, pcu, acu);
